Link Gdk.WidgetEvent to PopulateTotals method
Discovered that probably due to the RDWidget being a composite widget or built into a Table that it wont trigger on ordinary ButtonPress, EnterNotify, Motion, etc events/signals. Instead it emits the Gdk.WidgetEvents which is yielding the Gdk.Exposure event. So the array of dynamically generated RDWidgets are now suppose to signal/event with that. TODO change names