Fix and Extend DePopulateUI -> DePopulateRDTable
DePopulateUI method is ok. Though more logic was needed to capture issues where the New ScoreCard is made when a previous ScoreCard rdwidget table doesnt exist. (a condition reachable by not Deleting the old ScoreCard).

Therefore create DePopulateRDTable (to attempt the reverse of PopulateRDTable) to be called by DePopulateUI. Also to be initially called on every New ScoreCard creation. Obviously the very first ScoreCard of the session wont have all GTK Elements (and specifically the rdwidget) to DePopulate. However, this is swalllowed in a NullReferenceException and the progam is continued. Further GTK+ already handles widget elements not existing by NULLing them. So should be good