Linkage Presentation Layer to Business Logic and others
Mainly refactoring to ensure code that is in the backend is appropriately updated when user interacts on the GUI in regards to individual Rounds and the scoring of it. Additional presentation to business linkage is still outstanding (in particularly Fight Comment, Special Round Notations, and Fight Title EntryBox)

Additionally, smaller property-ize some of the class fields.
Remove the old stagnant Stetic GUI menubar from xmls
Toggle "Save ScoreCard" Sensitive property thru the *Populate* methods.